v3.4 Partial Map Class Compatibility
Beginning to convert the map to class based programming. 3 classes have been added for rooms with events, rooms with alt descriptions, and normal rooms. main.py and some of searchnevent.py has been reprogrammed to support this. map.py now has a function to create the room object and a function to save the changed variables and delete the object.